% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{create_matchmaking_rule_set}
\alias{create_matchmaking_rule_set}
\title{Creates a new rule set for FlexMatch matchmaking}
\usage{
create_matchmaking_rule_set(Name, RuleSetBody)
}
\arguments{
\item{Name}{[required] Unique identifier for a matchmaking rule set. A matchmaking configuration identifies the rule set it uses by this name value. (Note: The rule set name is different from the optional "name" field in the rule set body.)}

\item{RuleSetBody}{[required] Collection of matchmaking rules, formatted as a JSON string. Note that comments are not allowed in JSON, but most elements support a description field.}
}
\description{
Creates a new rule set for FlexMatch matchmaking. A rule set describes the type of match to create, such as the number and size of teams, and sets the parameters for acceptable player matches, such as minimum skill level or character type. A rule set is used by a MatchmakingConfiguration.
}
\details{
To create a matchmaking rule set, provide unique rule set name and the rule set body in JSON format. Rule sets must be defined in the same region as the matchmaking configuration they will be used with.

Since matchmaking rule sets cannot be edited, it is a good idea to check the rule set syntax using ValidateMatchmakingRuleSet before creating a new rule set.

\strong{Learn more}
\itemize{
\item \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/match-rulesets.html}{Build a Rule Set}
\item \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/match-configuration.html}{Design a Matchmaker}
\item \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/match-intro.html}{Matchmaking with FlexMatch}
}

\strong{Related operations}
\itemize{
\item CreateMatchmakingConfiguration
\item DescribeMatchmakingConfigurations
\item UpdateMatchmakingConfiguration
\item DeleteMatchmakingConfiguration
\item CreateMatchmakingRuleSet
\item DescribeMatchmakingRuleSets
\item ValidateMatchmakingRuleSet
\item DeleteMatchmakingRuleSet
}
}
\section{Accepted Parameters}{
\preformatted{create_matchmaking_rule_set(
  Name = "string",
  RuleSetBody = "string"
)
}
}

