% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{create_player_session}
\alias{create_player_session}
\title{Adds a player to a game session and creates a player session record}
\usage{
create_player_session(
  GameSessionId = "string",
  PlayerId = "string",
  PlayerData = "string"
)
}
\arguments{
\item{GameSessionId}{[required] Unique identifier for the game session to add a player to.}

\item{PlayerId}{[required] Unique identifier for a player. Player IDs are developer-defined.}

\item{PlayerData}{Developer-defined information related to a player. Amazon GameLift does not use this data, so it can be formatted as needed for use in the game.}
}
\description{
Adds a player to a game session and creates a player session record. Before a player can be added, a game session must have an \code{ACTIVE} status, have a creation policy of \code{ALLOW_ALL}, and have an open player slot. To add a group of players to a game session, use CreatePlayerSessions.
}
\details{
To create a player session, specify a game session ID, player ID, and optionally a string of player data. If successful, the player is added to the game session and a new PlayerSession object is returned. Player sessions cannot be updated.

\emph{Available in Amazon GameLift Local.}

Player-session-related operations include:
\itemize{
\item CreatePlayerSession
\item CreatePlayerSessions
\item DescribePlayerSessions
\item Game session placements
\item StartGameSessionPlacement
\item DescribeGameSessionPlacement
\item StopGameSessionPlacement
}
}
\examples{

}
