% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{create_player_sessions}
\alias{create_player_sessions}
\title{Adds a group of players to a game session}
\usage{
create_player_sessions(GameSessionId, PlayerIds, PlayerDataMap = NULL)
}
\arguments{
\item{GameSessionId}{Unique identifier for the game session to add players to.}

\item{PlayerIds}{List of unique identifiers for the players to be added.}

\item{PlayerDataMap}{Map of string pairs, each specifying a player ID and a set of developer-defined information related to the player. Amazon GameLift does not use this data, so it can be formatted as needed for use in the game. Player data strings for player IDs not included in the \code{PlayerIds} parameter are ignored.}
}
\description{
Adds a group of players to a game session. This action is useful with a team matching feature. Before players can be added, a game session must have an \code{ACTIVE} status, have a creation policy of \code{ALLOW_ALL}, and have an open player slot. To add a single player to a game session, use CreatePlayerSession.
}
\details{
To create player sessions, specify a game session ID, a list of player IDs, and optionally a set of player data strings. If successful, the players are added to the game session and a set of new PlayerSession objects is returned. Player sessions cannot be updated.

\emph{Available in Amazon GameLift Local.}

Player-session-related operations include:
\itemize{
\item CreatePlayerSession
\item CreatePlayerSessions
\item DescribePlayerSessions
\item Game session placements
\item StartGameSessionPlacement
\item DescribeGameSessionPlacement
\item StopGameSessionPlacement
}
}
\examples{

}
