% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{delete_build}
\alias{delete_build}
\title{Deletes a build}
\usage{
delete_build(BuildId)
}
\arguments{
\item{BuildId}{[required] Unique identifier for a build to delete.}
}
\description{
Deletes a build. This action permanently deletes the build record and any uploaded build files.
}
\details{
To delete a build, specify its ID. Deleting a build does not affect the status of any active fleets using the build, but you can no longer create new fleets with the deleted build.

Build-related operations include:
\itemize{
\item CreateBuild
\item ListBuilds
\item DescribeBuild
\item UpdateBuild
\item DeleteBuild
}
}
\section{Accepted Parameters}{
\preformatted{delete_build(
  BuildId = "string"
)
}
}

