% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{delete_matchmaking_configuration}
\alias{delete_matchmaking_configuration}
\title{Permanently removes a FlexMatch matchmaking configuration}
\usage{
delete_matchmaking_configuration(Name)
}
\arguments{
\item{Name}{[required] Unique identifier for a matchmaking configuration}
}
\description{
Permanently removes a FlexMatch matchmaking configuration. To delete, specify the configuration name. A matchmaking configuration cannot be deleted if it is being used in any active matchmaking tickets.
}
\details{
Operations related to match configurations and rule sets include:
\itemize{
\item CreateMatchmakingConfiguration
\item DescribeMatchmakingConfigurations
\item UpdateMatchmakingConfiguration
\item DeleteMatchmakingConfiguration
\item CreateMatchmakingRuleSet
\item DescribeMatchmakingRuleSets
\item ValidateMatchmakingRuleSet
}
}
\section{Accepted Parameters}{
\preformatted{delete_matchmaking_configuration(
  Name = "string"
)
}
}

