% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{delete_vpc_peering_connection}
\alias{delete_vpc_peering_connection}
\title{Removes a VPC peering connection}
\usage{
delete_vpc_peering_connection(FleetId, VpcPeeringConnectionId)
}
\arguments{
\item{FleetId}{[required] Unique identifier for a fleet. This value must match the fleet ID referenced in the VPC peering connection record.}

\item{VpcPeeringConnectionId}{[required] Unique identifier for a VPC peering connection. This value is included in the VpcPeeringConnection object, which can be retrieved by calling DescribeVpcPeeringConnections.}
}
\description{
Removes a VPC peering connection. To delete the connection, you must have a valid authorization for the VPC peering connection that you want to delete. You can check for an authorization by calling DescribeVpcPeeringAuthorizations or request a new one using CreateVpcPeeringAuthorization.
}
\details{
Once a valid authorization exists, call this operation from the AWS account that is used to manage the Amazon GameLift fleets. Identify the connection to delete by the connection ID and fleet ID. If successful, the connection is removed.

VPC peering connection operations include:
\itemize{
\item CreateVpcPeeringAuthorization
\item DescribeVpcPeeringAuthorizations
\item DeleteVpcPeeringAuthorization
\item CreateVpcPeeringConnection
\item DescribeVpcPeeringConnections
\item DeleteVpcPeeringConnection
}
}
\section{Accepted Parameters}{
\preformatted{delete_vpc_peering_connection(
  FleetId = "string",
  VpcPeeringConnectionId = "string"
)
}
}

