% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{describe_alias}
\alias{describe_alias}
\title{Retrieves properties for an alias}
\usage{
describe_alias(AliasId)
}
\arguments{
\item{AliasId}{Unique identifier for a fleet alias. Specify the alias you want to retrieve.}
}
\description{
Retrieves properties for an alias. This operation returns all alias metadata and settings. To get an alias's target fleet ID only, use \code{ResolveAlias}.
}
\details{
To get alias properties, specify the alias ID. If successful, the requested alias record is returned.

Alias-related operations include:
\itemize{
\item CreateAlias
\item ListAliases
\item DescribeAlias
\item UpdateAlias
\item DeleteAlias
\item ResolveAlias
}
}
\examples{

}
