% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{describe_build}
\alias{describe_build}
\title{Retrieves properties for a build}
\usage{
describe_build(BuildId)
}
\arguments{
\item{BuildId}{Unique identifier for a build to retrieve properties for.}
}
\description{
Retrieves properties for a build. To request a build record, specify a build ID. If successful, an object containing the build properties is returned.
}
\details{
Build-related operations include:
\itemize{
\item CreateBuild
\item ListBuilds
\item DescribeBuild
\item UpdateBuild
\item DeleteBuild
}
}
\examples{

}
