% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{describe_ec2_instance_limits}
\alias{describe_ec2_instance_limits}
\title{Retrieves the following information for the specified EC2 instance type:}
\usage{
describe_ec2_instance_limits(EC2InstanceType = NULL)
}
\arguments{
\item{EC2InstanceType}{Name of an EC2 instance type that is supported in Amazon GameLift. A fleet instance type determines the computing resources of each instance in the fleet, including CPU, memory, storage, and networking capacity. Amazon GameLift supports the following EC2 instance types. See \href{http://aws.amazon.com/ec2/instance-types/}{Amazon EC2 Instance Types} for detailed descriptions. Leave this parameter blank to retrieve limits for all types.}
}
\description{
Retrieves the following information for the specified EC2 instance type:
}
\details{
\itemize{
\item maximum number of instances allowed per AWS account (service limit)
\item current usage level for the AWS account
}

Service limits vary depending on region. Available regions for Amazon GameLift can be found in the AWS Management Console for Amazon GameLift (see the drop-down list in the upper right corner).

Fleet-related operations include:
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
\item Update fleets:
\item UpdateFleetAttributes
\item UpdateFleetCapacity
\item UpdateFleetPortSettings
\item UpdateRuntimeConfiguration
\item Manage fleet actions:
\item StartFleetActions
\item StopFleetActions
}
}
\examples{

}
