% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{describe_fleet_attributes}
\alias{describe_fleet_attributes}
\title{Retrieves fleet properties, including metadata, status, and configuration, for one or more fleets}
\usage{
describe_fleet_attributes(FleetIds = NULL, Limit = NULL,
  NextToken = NULL)
}
\arguments{
\item{FleetIds}{Unique identifier for a fleet(s) to retrieve attributes for. To request attributes for all fleets, leave this parameter empty.}

\item{Limit}{Maximum number of results to return. Use this parameter with \code{NextToken} to get results as a set of sequential pages. This parameter is ignored when the request specifies one or a list of fleet IDs.}

\item{NextToken}{Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value. This parameter is ignored when the request specifies one or a list of fleet IDs.}
}
\description{
Retrieves fleet properties, including metadata, status, and configuration, for one or more fleets. You can request attributes for all fleets, or specify a list of one or more fleet IDs. When requesting multiple fleets, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a FleetAttributes object is returned for each requested fleet ID. When specifying a list of fleet IDs, attribute objects are returned only for fleets that currently exist.
}
\details{
Some API actions may limit the number of fleet IDs allowed in one request. If a request exceeds this limit, the request fails and the error message includes the maximum allowed.
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
\item Update fleets:
\item UpdateFleetAttributes
\item UpdateFleetCapacity
\item UpdateFleetPortSettings
\item UpdateRuntimeConfiguration
\item Manage fleet actions:
\item StartFleetActions
\item StopFleetActions
}
}
\section{Accepted Parameters}{
\preformatted{describe_fleet_attributes(
  FleetIds = list(
    "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

