% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{describe_game_sessions}
\alias{describe_game_sessions}
\title{Retrieves a set of one or more game sessions}
\usage{
describe_game_sessions(FleetId = NULL, GameSessionId = NULL,
  AliasId = NULL, StatusFilter = NULL, Limit = NULL, NextToken = NULL)
}
\arguments{
\item{FleetId}{Unique identifier for a fleet to retrieve all game sessions for.}

\item{GameSessionId}{Unique identifier for the game session to retrieve. You can use either a \code{GameSessionId} or \code{GameSessionArn} value.}

\item{AliasId}{Unique identifier for an alias associated with the fleet to retrieve all game sessions for.}

\item{StatusFilter}{Game session status to filter results on. Possible game session statuses include \code{ACTIVE}, \code{TERMINATED}, \code{ACTIVATING}, and \code{TERMINATING} (the last two are transitory).}

\item{Limit}{Maximum number of results to return. Use this parameter with \code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves a set of one or more game sessions. Request a specific game session or request all game sessions on a fleet. Alternatively, use SearchGameSessions to request a set of active game sessions that are filtered by certain criteria. To retrieve protection policy settings for game sessions, use DescribeGameSessionDetails.
}
\details{
To get game sessions, specify one of the following: game session ID, fleet ID, or alias ID. You can filter this request by game session status. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, a GameSession object is returned for each game session matching the request.

\emph{Available in Amazon GameLift Local.}

Game-session-related operations include:
\itemize{
\item CreateGameSession
\item DescribeGameSessions
\item DescribeGameSessionDetails
\item SearchGameSessions
\item UpdateGameSession
\item GetGameSessionLogUrl
\item Game session placements
\item StartGameSessionPlacement
\item DescribeGameSessionPlacement
\item StopGameSessionPlacement
}
}
\examples{

}
