% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{describe_matchmaking_configurations}
\alias{describe_matchmaking_configurations}
\title{Retrieves the details of FlexMatch matchmaking configurations}
\usage{
describe_matchmaking_configurations(Names = NULL, RuleSetName = NULL,
  Limit = NULL, NextToken = NULL)
}
\arguments{
\item{Names}{Unique identifier for a matchmaking configuration(s) to retrieve. To request all existing configurations, leave this parameter empty.}

\item{RuleSetName}{Unique identifier for a matchmaking rule set. Use this parameter to retrieve all matchmaking configurations that use this rule set.}

\item{Limit}{Maximum number of results to return. Use this parameter with \code{NextToken} to get results as a set of sequential pages. This parameter is limited to 10.}

\item{NextToken}{Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves the details of FlexMatch matchmaking configurations. with this operation, you have the following options: (1) retrieve all existing configurations, (2) provide the names of one or more configurations to retrieve, or (3) retrieve all configurations that use a specified rule set name. When requesting multiple items, use the pagination parameters to retrieve results as a set of sequential pages. If successful, a configuration is returned for each requested name. When specifying a list of names, only configurations that currently exist are returned.
}
\details{
\itemize{
\item CreateMatchmakingConfiguration
\item DescribeMatchmakingConfigurations
\item UpdateMatchmakingConfiguration
\item DeleteMatchmakingConfiguration
\item CreateMatchmakingRuleSet
\item DescribeMatchmakingRuleSets
\item ValidateMatchmakingRuleSet
\item DeleteMatchmakingRuleSet
}
}
\section{Accepted Parameters}{
\preformatted{describe_matchmaking_configurations(
  Names = list(
    "string"
  ),
  RuleSetName = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

