% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{describe_player_sessions}
\alias{describe_player_sessions}
\title{Retrieves properties for one or more player sessions}
\usage{
describe_player_sessions(
  GameSessionId = "string",
  PlayerId = "string",
  PlayerSessionId = "string",
  PlayerSessionStatusFilter = "string",
  Limit = 123,
  NextToken = "string"
)
}
\arguments{
\item{GameSessionId}{Unique identifier for the game session to retrieve player sessions for.}

\item{PlayerId}{Unique identifier for a player to retrieve player sessions for.}

\item{PlayerSessionId}{Unique identifier for a player session to retrieve.}

\item{PlayerSessionStatusFilter}{Player session status to filter results on.

Possible player session statuses include the following:
\itemize{
\item \strong{RESERVED} -- The player session request has been received, but the player has not yet connected to the server process and/or been validated.
\item \strong{ACTIVE} -- The player has been validated by the server process and is currently connected.
\item \strong{COMPLETED} -- The player connection has been dropped.
\item \strong{TIMEDOUT} -- A player session request was received, but the player did not connect and/or was not validated within the timeout limit (60 seconds).
}}

\item{Limit}{Maximum number of results to return. Use this parameter with \code{NextToken} to get results as a set of sequential pages. If a player session ID is specified, this parameter is ignored.}

\item{NextToken}{Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value. If a player session ID is specified, this parameter is ignored.}
}
\description{
Retrieves properties for one or more player sessions. This action can be used in several ways: (1) provide a \code{PlayerSessionId} to request properties for a specific player session; (2) provide a \code{GameSessionId} to request properties for all player sessions in the specified game session; (3) provide a \code{PlayerId} to request properties for all player sessions of a specified player.
}
\details{
To get game session record(s), specify only one of the following: a player session ID, a game session ID, or a player ID. You can filter this request by player session status. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, a PlayerSession object is returned for each session matching the request.

\emph{Available in Amazon GameLift Local.}

Player-session-related operations include:
\itemize{
\item CreatePlayerSession
\item CreatePlayerSessions
\item DescribePlayerSessions
\item Game session placements
\item StartGameSessionPlacement
\item DescribeGameSessionPlacement
\item StopGameSessionPlacement
}
}
\examples{

}
