% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{describe_scaling_policies}
\alias{describe_scaling_policies}
\title{Retrieves all scaling policies applied to a fleet}
\usage{
describe_scaling_policies(FleetId, StatusFilter = NULL, Limit = NULL,
  NextToken = NULL)
}
\arguments{
\item{FleetId}{[required] Unique identifier for a fleet to retrieve scaling policies for.}

\item{StatusFilter}{Scaling policy status to filter results on. A scaling policy is only in force when in an \code{ACTIVE} status.
\itemize{
\item \strong{ACTIVE} -- The scaling policy is currently in force.
\item \strong{UPDATEREQUESTED} -- A request to update the scaling policy has been received.
\item \strong{UPDATING} -- A change is being made to the scaling policy.
\item \strong{DELETEREQUESTED} -- A request to delete the scaling policy has been received.
\item \strong{DELETING} -- The scaling policy is being deleted.
\item \strong{DELETED} -- The scaling policy has been deleted.
\item \strong{ERROR} -- An error occurred in creating the policy. It should be removed and recreated.
}}

\item{Limit}{Maximum number of results to return. Use this parameter with \code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves all scaling policies applied to a fleet.
}
\details{
To get a fleet's scaling policies, specify the fleet ID. You can filter this request by policy status, such as to retrieve only active scaling policies. Use the pagination parameters to retrieve results as a set of sequential pages. If successful, set of ScalingPolicy objects is returned for the fleet.

A fleet may have all of its scaling policies suspended (StopFleetActions). This action does not affect the status of the scaling policies, which remains ACTIVE. To see whether a fleet's scaling policies are in force or suspended, call DescribeFleetAttributes and check the stopped actions.

Operations related to fleet capacity scaling include:
\itemize{
\item DescribeFleetCapacity
\item UpdateFleetCapacity
\item DescribeEC2InstanceLimits
\item Manage scaling policies:
\item PutScalingPolicy (auto-scaling)
\item DescribeScalingPolicies (auto-scaling)
\item DeleteScalingPolicy (auto-scaling)
\item Manage fleet actions:
\item StartFleetActions
\item StopFleetActions
}
}
\section{Accepted Parameters}{
\preformatted{describe_scaling_policies(
  FleetId = "string",
  StatusFilter = "ACTIVE"|"UPDATE_REQUESTED"|"UPDATING"|"DELETE_REQUESTED"|"DELETING"|"DELETED"|"ERROR",
  Limit = 123,
  NextToken = "string"
)
}
}

