% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{describe_vpc_peering_authorizations}
\alias{describe_vpc_peering_authorizations}
\title{Retrieves valid VPC peering authorizations that are pending for the AWS account}
\usage{
describe_vpc_peering_authorizations()
}
\description{
Retrieves valid VPC peering authorizations that are pending for the AWS account. This operation returns all VPC peering authorizations and requests for peering. This includes those initiated and received by this account.
}
\details{
VPC peering connection operations include:
\itemize{
\item CreateVpcPeeringAuthorization
\item DescribeVpcPeeringAuthorizations
\item DeleteVpcPeeringAuthorization
\item CreateVpcPeeringConnection
\item DescribeVpcPeeringConnections
\item DeleteVpcPeeringConnection
}
}
\examples{

}
