% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{list_fleets}
\alias{list_fleets}
\title{Retrieves a collection of fleet records for this AWS account}
\usage{
list_fleets(BuildId = NULL, Limit = NULL, NextToken = NULL)
}
\arguments{
\item{BuildId}{Unique identifier for a build to return fleets for. Use this parameter to return only fleets using the specified build. To retrieve all fleets, leave this parameter empty.}

\item{Limit}{Maximum number of results to return. Use this parameter with \code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{Token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this action. To start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves a collection of fleet records for this AWS account. You can filter the result set by build ID. Use the pagination parameters to retrieve results in sequential pages.
}
\details{
Fleet records are not listed in any particular order.

Fleet-related operations include:
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
\item Update fleets:
\item UpdateFleetAttributes
\item UpdateFleetCapacity
\item UpdateFleetPortSettings
\item UpdateRuntimeConfiguration
\item Manage fleet actions:
\item StartFleetActions
\item StopFleetActions
}
}
\examples{

}
