% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{request_upload_credentials}
\alias{request_upload_credentials}
\title{Retrieves a fresh set of credentials for use when uploading a new set of game build files to Amazon GameLift's Amazon S3}
\usage{
request_upload_credentials(BuildId)
}
\arguments{
\item{BuildId}{Unique identifier for a build to get credentials for.}
}
\description{
Retrieves a fresh set of credentials for use when uploading a new set of game build files to Amazon GameLift's Amazon S3. This is done as part of the build creation process; see CreateBuild.
}
\details{
To request new credentials, specify the build ID as returned with an initial \code{CreateBuild} request. If successful, a new set of credentials are returned, along with the S3 storage location associated with the build ID.
}
\examples{

}
