% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{resolve_alias}
\alias{resolve_alias}
\title{Retrieves the fleet ID that a specified alias is currently pointing to}
\usage{
resolve_alias(AliasId)
}
\arguments{
\item{AliasId}{[required] Unique identifier for the alias you want to resolve.}
}
\description{
Retrieves the fleet ID that a specified alias is currently pointing to.
}
\details{
Alias-related operations include:
\itemize{
\item CreateAlias
\item ListAliases
\item DescribeAlias
\item UpdateAlias
\item DeleteAlias
\item ResolveAlias
}
}
\section{Accepted Parameters}{
\preformatted{resolve_alias(
  AliasId = "string"
)
}
}

