% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{start_fleet_actions}
\alias{start_fleet_actions}
\title{Resumes activity on a fleet that was suspended with StopFleetActions}
\usage{
start_fleet_actions(FleetId, Actions)
}
\arguments{
\item{FleetId}{Unique identifier for a fleet}

\item{Actions}{List of actions to restart on the fleet.}
}
\description{
Resumes activity on a fleet that was suspended with StopFleetActions. Currently, this operation is used to restart a fleet's auto-scaling activity.
}
\details{
To start fleet actions, specify the fleet ID and the type of actions to restart. When auto-scaling fleet actions are restarted, Amazon GameLift once again initiates scaling events as triggered by the fleet's scaling policies. If actions on the fleet were never stopped, this operation will have no effect. You can view a fleet's stopped actions using DescribeFleetAttributes.

Operations related to fleet capacity scaling include:
\itemize{
\item DescribeFleetCapacity
\item UpdateFleetCapacity
\item DescribeEC2InstanceLimits
\item Manage scaling policies:
\item PutScalingPolicy (auto-scaling)
\item DescribeScalingPolicies (auto-scaling)
\item DeleteScalingPolicy (auto-scaling)
\item Manage fleet actions:
\item StartFleetActions
\item StopFleetActions
}
}
\examples{

}
