% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{start_game_session_placement}
\alias{start_game_session_placement}
\title{Places a request for a new game session in a queue (see CreateGameSessionQueue)}
\usage{
start_game_session_placement(
  PlacementId = "string",
  GameSessionQueueName = "string",
  GameProperties = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  MaximumPlayerSessionCount = 123,
  GameSessionName = "string",
  PlayerLatencies = list(
    list(
      PlayerId = "string",
      RegionIdentifier = "string",
      LatencyInMilliseconds = 123.0
    )
  ),
  DesiredPlayerSessions = list(
    list(
      PlayerId = "string",
      PlayerData = "string"
    )
  ),
  GameSessionData = "string"
)
}
\arguments{
\item{PlacementId}{[required] Unique identifier to assign to the new game session placement. This value is developer-defined. The value must be unique across all regions and cannot be reused unless you are resubmitting a canceled or timed-out placement request.}

\item{GameSessionQueueName}{[required] Name of the queue to use to place the new game session.}

\item{GameProperties}{Set of custom properties for a game session, formatted as key:value pairs. These properties are passed to a game server process in the GameSession object with a request to start a new game session (see \href{http://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).}

\item{MaximumPlayerSessionCount}{[required] Maximum number of players that can be connected simultaneously to the game session.}

\item{GameSessionName}{Descriptive label that is associated with a game session. Session names do not need to be unique.}

\item{PlayerLatencies}{Set of values, expressed in milliseconds, indicating the amount of latency that a player experiences when connected to AWS regions. This information is used to try to place the new game session where it can offer the best possible gameplay experience for the players.}

\item{DesiredPlayerSessions}{Set of information on each player to create a player session for.}

\item{GameSessionData}{Set of custom game session properties, formatted as a single string value. This data is passed to a game server process in the GameSession object with a request to start a new game session (see \href{http://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).}
}
\description{
Places a request for a new game session in a queue (see CreateGameSessionQueue). When processing a placement request, Amazon GameLift searches for available resources on the queue's destinations, scanning each until it finds resources or the placement request times out.
}
\details{
A game session placement request can also request player sessions. When a new game session is successfully created, Amazon GameLift creates a player session for each player included in the request.

When placing a game session, by default Amazon GameLift tries each fleet in the order they are listed in the queue configuration. Ideally, a queue's destinations are listed in preference order.

Alternatively, when requesting a game session with players, you can also provide latency data for each player in relevant regions. Latency data indicates the performance lag a player experiences when connected to a fleet in the region. Amazon GameLift uses latency data to reorder the list of destinations to place the game session in a region with minimal lag. If latency data is provided for multiple players, Amazon GameLift calculates each region's average lag for all players and reorders to get the best game play across all players.

To place a new game session request, specify the following:
\itemize{
\item The queue name and a set of game session properties and settings
\item A unique ID (such as a UUID) for the placement. You use this ID to track the status of the placement request
\item (Optional) A set of IDs and player data for each player you want to join to the new game session
\item Latency data for all players (if you want to optimize game play for the players)
}

If successful, a new game session placement is created.

To track the status of a placement request, call DescribeGameSessionPlacement and check the request's status. If the status is \code{FULFILLED}, a new game session has been created and a game session ARN and region are referenced. If the placement request times out, you can resubmit the request or retry it with a different queue.

Game-session-related operations include:
\itemize{
\item CreateGameSession
\item DescribeGameSessions
\item DescribeGameSessionDetails
\item SearchGameSessions
\item UpdateGameSession
\item GetGameSessionLogUrl
\item Game session placements
\item StartGameSessionPlacement
\item DescribeGameSessionPlacement
\item StopGameSessionPlacement
}
}
\examples{

}
