% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{start_matchmaking}
\alias{start_matchmaking}
\title{Uses FlexMatch to create a game match for a group of players based on custom matchmaking rules, and starts a new game for the matched players}
\usage{
start_matchmaking(
  TicketId = "string",
  ConfigurationName = "string",
  Players = list(
    list(
      PlayerId = "string",
      PlayerAttributes = list(
        list(
          S = "string",
          N = 123.0,
          SL = list(
            "string"
          ),
          SDM = list(
            123.0
          )
        )
      ),
      Team = "string",
      LatencyInMs = list(
        123
      )
    )
  )
)
}
\arguments{
\item{TicketId}{Unique identifier for a matchmaking ticket. If no ticket ID is specified here, Amazon GameLift will generate one in the form of a UUID. Use this identifier to track the matchmaking ticket status and retrieve match results.}

\item{ConfigurationName}{[required] Name of the matchmaking configuration to use for this request. Matchmaking configurations must exist in the same region as this request.}

\item{Players}{[required] Information on each player to be matched. This information must include a player ID, and may contain player attributes and latency data to be used in the matchmaking process. After a successful match, \code{Player} objects contain the name of the team the player is assigned to.}
}
\description{
Uses FlexMatch to create a game match for a group of players based on custom matchmaking rules, and starts a new game for the matched players. Each matchmaking request specifies the type of match to build (team configuration, rules for an acceptable match, etc.). The request also specifies the players to find a match for and where to host the new game session for optimal performance. A matchmaking request might start with a single player or a group of players who want to play together. FlexMatch finds additional players as needed to fill the match. Match type, rules, and the queue used to place a new game session are defined in a \code{MatchmakingConfiguration}. For complete information on setting up and using FlexMatch, see the topic \href{http://docs.aws.amazon.com/gamelift/latest/developerguide/match-intro.html}{Adding FlexMatch to Your Game}.
}
\details{
To start matchmaking, provide a unique ticket ID, specify a matchmaking configuration, and include the players to be matched. You must also include a set of player attributes relevant for the matchmaking configuration. If successful, a matchmaking ticket is returned with status set to \code{QUEUED}. Track the status of the ticket to respond as needed and acquire game session connection information for successfully completed matches.

\strong{Tracking ticket status} -- A couple of options are available for tracking the status of matchmaking requests:
\itemize{
\item Polling -- Call \code{DescribeMatchmaking}. This operation returns the full ticket object, including current status and (for completed tickets) game session connection info. We recommend polling no more than once every 10 seconds.
\item Notifications -- Get event notifications for changes in ticket status using Amazon Simple Notification Service (SNS). Notifications are easy to set up (see CreateMatchmakingConfiguration) and typically deliver match status changes faster and more efficiently than polling. We recommend that you use polling to back up to notifications (since delivery is not guaranteed) and call \code{DescribeMatchmaking} only when notifications are not received within 30 seconds.
}

\strong{Processing a matchmaking request} -- FlexMatch handles a matchmaking request as follows:
\enumerate{
\item Your client code submits a \code{StartMatchmaking} request for one or more players and tracks the status of the request ticket.
\item FlexMatch uses this ticket and others in process to build an acceptable match. When a potential match is identified, all tickets in the proposed match are advanced to the next status.
\item If the match requires player acceptance (set in the matchmaking configuration), the tickets move into status \code{REQUIRES_ACCEPTANCE}. This status triggers your client code to solicit acceptance from all players in every ticket involved in the match, and then call AcceptMatch for each player. If any player rejects or fails to accept the match before a specified timeout, the proposed match is dropped (see \code{AcceptMatch} for more details).
\item Once a match is proposed and accepted, the matchmaking tickets move into status \code{PLACING}. FlexMatch locates resources for a new game session using the game session queue (set in the matchmaking configuration) and creates the game session based on the match data.
\item When the match is successfully placed, the matchmaking tickets move into \code{COMPLETED} status. Connection information (including game session endpoint and player session) is added to the matchmaking tickets. Matched players can use the connection information to join the game.
}

Matchmaking-related operations include:
\itemize{
\item StartMatchmaking
\item DescribeMatchmaking
\item StopMatchmaking
\item AcceptMatch
\item StartMatchBackfill
}
}
\examples{

}
