% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{stop_fleet_actions}
\alias{stop_fleet_actions}
\title{Suspends activity on a fleet}
\usage{
stop_fleet_actions(FleetId, Actions)
}
\arguments{
\item{FleetId}{Unique identifier for a fleet}

\item{Actions}{List of actions to suspend on the fleet.}
}
\description{
Suspends activity on a fleet. Currently, this operation is used to stop a fleet's auto-scaling activity. It is used to temporarily stop scaling events triggered by the fleet's scaling policies. The policies can be retained and auto-scaling activity can be restarted using StartFleetActions. You can view a fleet's stopped actions using DescribeFleetAttributes.
}
\details{
To stop fleet actions, specify the fleet ID and the type of actions to suspend. When auto-scaling fleet actions are stopped, Amazon GameLift no longer initiates scaling events except to maintain the fleet's desired instances setting (FleetCapacity. Changes to the fleet's capacity must be done manually using UpdateFleetCapacity.
}
\examples{

}
