% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{stop_game_session_placement}
\alias{stop_game_session_placement}
\title{Cancels a game session placement that is in \code{PENDING} status}
\usage{
stop_game_session_placement(
  PlacementId = "string"
)
}
\arguments{
\item{PlacementId}{[required] Unique identifier for a game session placement to cancel.}
}
\description{
Cancels a game session placement that is in \code{PENDING} status. To stop a placement, provide the placement ID values. If successful, the placement is moved to \code{CANCELLED} status.
}
\details{
Game-session-related operations include:
\itemize{
\item CreateGameSession
\item DescribeGameSessions
\item DescribeGameSessionDetails
\item SearchGameSessions
\item UpdateGameSession
\item GetGameSessionLogUrl
\item Game session placements
\item StartGameSessionPlacement
\item DescribeGameSessionPlacement
\item StopGameSessionPlacement
}
}
\examples{

}
