% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{stop_matchmaking}
\alias{stop_matchmaking}
\title{Cancels a matchmaking ticket that is currently being processed}
\usage{
stop_matchmaking(TicketId)
}
\arguments{
\item{TicketId}{[required] Unique identifier for a matchmaking ticket.}
}
\description{
Cancels a matchmaking ticket that is currently being processed. To stop the matchmaking operation, specify the ticket ID. If successful, work on the ticket is stopped, and the ticket status is changed to \code{CANCELLED}.
}
\details{
Matchmaking-related operations include:
\itemize{
\item StartMatchmaking
\item DescribeMatchmaking
\item StopMatchmaking
\item AcceptMatch
\item StartMatchBackfill
}
}
\section{Accepted Parameters}{
\preformatted{stop_matchmaking(
  TicketId = "string"
)
}
}

