% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{update_build}
\alias{update_build}
\title{Updates metadata in a build record, including the build name and version}
\usage{
update_build(BuildId, Name = NULL, Version = NULL)
}
\arguments{
\item{BuildId}{[required] Unique identifier for a build to update.}

\item{Name}{Descriptive label that is associated with a build. Build names do not need to be unique.}

\item{Version}{Version that is associated with this build. Version strings do not need to be unique.}
}
\description{
Updates metadata in a build record, including the build name and version. To update the metadata, specify the build ID to update and provide the new values. If successful, a build object containing the updated metadata is returned.
}
\details{
\itemize{
\item CreateBuild
\item ListBuilds
\item DescribeBuild
\item UpdateBuild
\item DeleteBuild
}
}
\section{Accepted Parameters}{
\preformatted{update_build(
  BuildId = "string",
  Name = "string",
  Version = "string"
)
}
}

