% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{update_fleet_attributes}
\alias{update_fleet_attributes}
\title{Updates fleet properties, including name and description, for a fleet}
\usage{
update_fleet_attributes(FleetId, Name = NULL, Description = NULL,
  NewGameSessionProtectionPolicy = NULL,
  ResourceCreationLimitPolicy = NULL, MetricGroups = NULL)
}
\arguments{
\item{FleetId}{[required] Unique identifier for a fleet to update attribute metadata for.}

\item{Name}{Descriptive label that is associated with a fleet. Fleet names do not need to be unique.}

\item{Description}{Human-readable description of a fleet.}

\item{NewGameSessionProtectionPolicy}{Game session protection policy to apply to all new instances created in this fleet. Instances that already exist are not affected. You can set protection for individual instances using UpdateGameSession.
\itemize{
\item \strong{NoProtection} -- The game session can be terminated during a scale-down event.
\item \strong{FullProtection} -- If the game session is in an \code{ACTIVE} status, it cannot be terminated during a scale-down event.
}}

\item{ResourceCreationLimitPolicy}{Policy that limits the number of game sessions an individual player can create over a span of time.}

\item{MetricGroups}{Names of metric groups to include this fleet in. Amazon CloudWatch uses a fleet metric group is to aggregate metrics from multiple fleets. Use an existing metric group name to add this fleet to the group. Or use a new name to create a new metric group. A fleet can only be included in one metric group at a time.}
}
\description{
Updates fleet properties, including name and description, for a fleet. To update metadata, specify the fleet ID and the property values that you want to change. If successful, the fleet ID for the updated fleet is returned.
}
\details{
Fleet-related operations include:
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
\item Update fleets:
\item UpdateFleetAttributes
\item UpdateFleetCapacity
\item UpdateFleetPortSettings
\item UpdateRuntimeConfiguration
\item Manage fleet actions:
\item StartFleetActions
\item StopFleetActions
}
}
\section{Accepted Parameters}{
\preformatted{update_fleet_attributes(
  FleetId = "string",
  Name = "string",
  Description = "string",
  NewGameSessionProtectionPolicy = "NoProtection"|"FullProtection",
  ResourceCreationLimitPolicy = list(
    NewGameSessionsPerCreator = 123,
    PolicyPeriodInMinutes = 123
  ),
  MetricGroups = list(
    "string"
  )
)
}
}

