% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{update_fleet_port_settings}
\alias{update_fleet_port_settings}
\title{Updates port settings for a fleet}
\usage{
update_fleet_port_settings(FleetId,
  InboundPermissionAuthorizations = NULL,
  InboundPermissionRevocations = NULL)
}
\arguments{
\item{FleetId}{[required] Unique identifier for a fleet to update port settings for.}

\item{InboundPermissionAuthorizations}{Collection of port settings to be added to the fleet record.}

\item{InboundPermissionRevocations}{Collection of port settings to be removed from the fleet record.}
}
\description{
Updates port settings for a fleet. To update settings, specify the fleet ID to be updated and list the permissions you want to update. List the permissions you want to add in \code{InboundPermissionAuthorizations}, and permissions you want to remove in \code{InboundPermissionRevocations}. Permissions to be removed must match existing fleet permissions. If successful, the fleet ID for the updated fleet is returned.
}
\details{
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
\item Update fleets:
\item UpdateFleetAttributes
\item UpdateFleetCapacity
\item UpdateFleetPortSettings
\item UpdateRuntimeConfiguration
\item Manage fleet actions:
\item StartFleetActions
\item StopFleetActions
}
}
\section{Accepted Parameters}{
\preformatted{update_fleet_port_settings(
  FleetId = "string",
  InboundPermissionAuthorizations = list(
    list(
      FromPort = 123,
      ToPort = 123,
      IpRange = "string",
      Protocol = "TCP"|"UDP"
    )
  ),
  InboundPermissionRevocations = list(
    list(
      FromPort = 123,
      ToPort = 123,
      IpRange = "string",
      Protocol = "TCP"|"UDP"
    )
  )
)
}
}

