% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.gamelift_operations.R
\name{validate_matchmaking_rule_set}
\alias{validate_matchmaking_rule_set}
\title{Validates the syntax of a matchmaking rule or rule set}
\usage{
validate_matchmaking_rule_set(
  RuleSetBody = "string"
)
}
\arguments{
\item{RuleSetBody}{[required] Collection of matchmaking rules to validate, formatted as a JSON string.}
}
\description{
Validates the syntax of a matchmaking rule or rule set. This operation checks that the rule set uses syntactically correct JSON and that it conforms to allowed property expressions. To validate syntax, provide a rule set string.
}
\details{
Operations related to match configurations and rule sets include:
\itemize{
\item CreateMatchmakingConfiguration
\item DescribeMatchmakingConfigurations
\item UpdateMatchmakingConfiguration
\item DeleteMatchmakingConfiguration
\item CreateMatchmakingRuleSet
\item DescribeMatchmakingRuleSets
\item ValidateMatchmakingRuleSet
}
}
\examples{

}
